HMW-393 Moved the SVG workaround to the utils file.
